+++ lib/rapid-vaults/api.rb
@@ -10,6 +10,13 @@ class RapidVaults::API
 
   # parse api options; this is mostly here for unit testing
   def self.parse(settings)
+    # validate args
+    if %i[encrypt decrypt].include?(settings[:action])
+      raise 'no file specified for encryption or decryption' if !settings.key?(:file)
+    end
+
+    raise 'input password cannot be empty' if settings.key?(:pw) && settings[:pw].empty?
+
     # establish settings for api and denote using api
     settings.merge({ ui: :api })
   end